+++ __init__.py
@@ -39,7 +39,7 @@ bl_info = {
     "blender": (2, 80, 0),
     "location": "View3D > Shift + Q key",
     "description": "Menu of material tools (assign, select..) in the 3D View",
-    "wiki_url": "https://docs.blender.org/manual/en/2.83/addons/materials/material_utils.html",
+    "doc_url": "{BLENDER_MANUAL_URL}/addons/materials/material_utils.html",
     "category": "Material"
 }
 
@@ -209,17 +209,6 @@ def unregister():
         if kmi.idname == "wm.call_menu":
             if kmi.properties.name == VIEW3D_MT_materialutilities_main.bl_idname:
                 km.keymap_items.remove(kmi)
-
-
-    """
-    if kc:
-        km = kc.keymaps["3D View"]
-        for kmi in km.keymap_items:
-            if kmi.idname == 'wm.call_menu':
-                if kmi.properties.name == VIEW3D_MT_materialutilities_main.bl_idname:
-                    km.keymap_items.remove(kmi)
-                    break
-    """
 
     mu_classes_unregister()